requirements.txt
Django==2.2.10
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0

